spec cleaned linear systems folder
