--- Labs/Vol1A/LinearSystems/oldlab/test.py
@@ -1,4 +1,0 @@
-def test(n):
-	t = 0
-	for i in xrange(n):
-    	t += sum(xrange(i))
