+++ Labs/Vol1A/LinearSystems/solutions.py
@@ -177,3 +177,61 @@ def ref_with_swaps(A):
     if m > n:
         A[n:] = 0
     return A
+
+def ref2(A):
+    """Alternate REF using an outer product. Fast, but not very intuitive."""
+    for i in xrange(A.shape[0]):
+        A[i+1:,i:] -= np.outer(A[i+1:,i]/A[i,i], A[i,i:])
+
+def lu2(A):
+    """Alternative LU decomposition using an outer product."""
+    U = A.copy()
+    L = np.eye(A.shape[0])
+    for i in xrange(A.shape[0]-1):
+        L[i+1:,i] = U[i+1:,i] / U[i,i]
+        U[i+1:,i:] -= np.outer(L[i+1:,i], U[i,i:])
+    return L, U
+
+def lu_inplace(A):
+    """Compute the LU decomposition of the square matrix A *IN PLACE*."""
+    for j in xrange(A.shape[0]-1):
+        for i in xrange(j+1, A.shape[0]):
+            A[i,j] /= A[j,j]
+            A[i,j+1:] -= A[i,j] * A[j,j+1:]
+
+def lu_det(A):
+    """Compute det(A) using the LU decomposition (via la.lu_factor())."""
+    lu, piv = la.lu_factor(A)
+
+    # Determine if there were an even or odd number of row swaps.
+    s = (piv != np.arange(A.shape[0])).sum() % 2
+    return ((-1)**s) * lu.diagonal().prod()
+
+def cholesky(A):
+    L = np.zeros_like(A)
+    for i in xrange(A.shape[0]):
+        for j in xrange(i):
+            L[i,j]=(A[i,j] - np.inner(L[i,:j], L[j,:j])) / L[j,j]
+        sl = L[i,:i]
+        L[i,i] = sqrt(A[i,i] - np.inner(sl, sl))
+    return L
+
+def cholesky_inplace(A):
+    for i in xrange(A.shape[0]):
+        A[i,i+1:] = 0.
+        for j in range(i):
+            A[i,j] = (A[i,j] - np.inner(A[i,:j],A[j,:j])) / A[j,j]
+        sl = A[i,:i]
+        A[i,i] = sqrt(A[i,i] - np.inner(sl, sl))
+
+def cholesky_solve(A, B):
+    for j in xrange(A.shape[0]):
+        B[j] /= A[j,j]
+        for i in xrange(j+1, A.shape[0]):
+            B[i] -= A[i,j] * B[j]
+    for j in xrange(A.shape[0]-1, -1, -1):
+        B[j] /= A[j,j]
+        for i in xrange(j):
+            B[i] -= A[j,i] * B[j]
+
+
